Export versions functions from commons to their own module
