+++ modules/commons.py
@@ -2,13 +2,7 @@
 Support for os, re, semver, version from packaging and other modules which written for avoiding repetitive code.
 """
 import os
-import re
-from packaging import version
-from modules import run_command, constants, docker_commands, os_type, os_release, supported
-VM_CLASS_HIERARCHY = 'VM.class_hierarchy'
-GC_CLASS_HISTOGRAM = 'GC.class_histogram'
-HELP = 'help'
-JDK_MINIMUM_VERSION = '10.0.0'
+from modules import run_command, constants, docker_commands, os_type, os_release
 
 
 def graph_start(cve, vol_graph):
@@ -73,84 +67,3 @@ def file_content(file_path, debug, container_name):
                 if content:
                     content = content.split('\n')[:constants.END]
     return content
-
-
-def re_start_of_version(full_version):
-    """Returns the start of a version using regex."""
-    start_of_version = re.search(r'\d*\.\d*', full_version)
-    if start_of_version:
-        return start_of_version.group()
-
-
-def check_patched_version(version_type, checked_version, patched_versions):
-    """This function checks if the version is affected according to patched versions which sorted in an ascending order
-     list."""
-    affected = False
-    print(constants.FULL_QUESTION_MESSAGE.format(f'Is {version_type} version affected?'))
-    if patched_versions[constants.START] in checked_version:
-        print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'Your {version_type} version which is: {checked_version} has '
-                                                        f'the patched version which is: '
-                                                        f'{patched_versions[constants.START]}'))
-    elif version.parse(checked_version) < version.parse(patched_versions[constants.START]):
-        print(constants.FULL_NEGATIVE_RESULT_MESSAGE.format('Yes'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'The lowest patched version is: '
-                                                        f'{patched_versions[constants.START]}\nYour {version_type}'
-                                                        f' version which is: {checked_version}, is affected'))
-        affected = True
-    elif version.parse(checked_version) > version.parse(patched_versions[constants.END]) or \
-            patched_versions[constants.END] in checked_version:
-        print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'The highest patched version is: '
-                                                        f'{patched_versions[constants.END]}\nYour {version_type}'
-                                                        f' version is: {checked_version}'))
-    else:
-        for patched_version in patched_versions[constants.FIRST:]:
-            start_of_checked_version = re_start_of_version(checked_version)
-            start_of_patched_version = re_start_of_version(patched_version)
-            if version.parse(start_of_checked_version) < version.parse(start_of_patched_version):
-                print(constants.FULL_NEGATIVE_RESULT_MESSAGE.format('Yes'))
-                print(constants.FULL_EXPLANATION_MESSAGE.format(f'Your {version_type} version which is: '
-                                                                f'{checked_version} is not patched'))
-                affected = True
-                break
-            if patched_version.startswith(start_of_checked_version):
-                if version.parse(checked_version) < version.parse(patched_version):
-                    print(constants.FULL_NEGATIVE_RESULT_MESSAGE.format('Yes'))
-                    print(constants.FULL_EXPLANATION_MESSAGE.format(f'The lowest patched version is: {patched_version}'
-                                                                    f'\nYour {version_type} version is: '
-                                                                    f'{checked_version}, is affected'))
-                    affected = True
-                    break
-                if patched_version in checked_version:
-                    print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-                    print(constants.FULL_EXPLANATION_MESSAGE.format(f'Your {version_type} version which is: '
-                                                                    f'{checked_version} has the patched version which '
-                                                                    f'is: {patched_versions}, is not affected'))
-                else:
-                    print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-                    print(constants.FULL_EXPLANATION_MESSAGE.format(f'The lowest patched version is: {patched_version}'
-                                                                    f'\nYour {version_type} version which is: '
-                                                                    f'{checked_version}, is not affected'))
-                    break
-    return affected
-
-
-def compare_versions(fixed_version, host_version, package_name):
-    """This function compares between the fixed version and the host's version."""
-    affected = False
-    print(constants.FULL_QUESTION_MESSAGE.format(f'Is {package_name} version affected?'))
-    if version.parse(fixed_version) < version.parse(host_version):
-        print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'The {package_name} patched versions is: {fixed_version}\nYour'
-                                                        f' versions which is: {host_version}, is not affected'))
-    elif version.parse(fixed_version) == version.parse(host_version):
-        print(constants.FULL_POSITIVE_RESULT_MESSAGE.format('No'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'Your system is not affected because it has the {package_name}'
-                                                        f' patched version which is: {fixed_version}'))
-    else:
-        print(constants.FULL_NEGATIVE_RESULT_MESSAGE.format('Yes'))
-        print(constants.FULL_EXPLANATION_MESSAGE.format(f'The {package_name} patched versions is: {fixed_version}\nYour'
-                                                        f' versions which is: {host_version}, is affected'))
-        affected = True
-    return affected
